Fix doubled records in bootstrap-select list
In appointment edit form I have fixed supervisors list.
Another two select controls in that form are waiting for changes to be made: physiotherapists and disease.
Also no solution for empty list passing to database e.g. when no diseases selected it shows previous state with selection.
When done it needs to be served to another controls in all forms.